Move display/xml_parser.js in shared to use it in worker
