+++ src/shared/xml_parser.js
